Start middleware implementation
