+++ middleware.ts
@@ -1,1 +1,8 @@
-export { auth as middleware } from "@/auth"
+import NextAuth from "next-auth";
+import { authConfig } from "./auth.config";
+
+export default NextAuth(authConfig).auth;
+
+export const config = {
+    matcher: ['/about/:path*', '/dashboard/:path*'],
+}
